Feat(LOCAT-BE-37): separate class and add interface
