--- src/main/java/com/locat/api/domain/user/entity/Platform.java
@@ -1,5 +1,0 @@
-package com.locat.api.domain.user.entity;
-
-public enum Platform {
-    ios, android
-}
