+++ src/main/java/com/locat/api/global/notification/NotificationServiceImpl.java
@@ -16,53 +16,7 @@ public class NotificationServiceImpl implements NotificationService {
     @Value("${service.aws.sns.topic-arn}")
     private String topicArn;
 
-    @Value("${service.aws.sns.platform-application-arn.ios}")
-    private String iosArn;
-
-    @Value("${service.aws.sns.platform-application-arn.android}")
-    private String androidArn;
     private final UserRepository userRepository;
-
-    public String createPlatformEndpoint(String token, String platform) {
-        String platformApplicationArn = getPlatformApplicationArn(platform);
-        CreatePlatformEndpointRequest request = CreatePlatformEndpointRequest.builder()
-                .token(token)
-                .platformApplicationArn(platformApplicationArn)
-                .build();
-
-        try {
-            CreatePlatformEndpointResponse response = this.snsClient.createPlatformEndpoint(request);
-            return response.endpointArn();
-        } catch (SnsException e) {
-            throw new RuntimeException("Failed to create platform endpoint", e); // Exception 추가 필요
-        }
-    }
-
-    public String subscribeEndpointToTopic(String endpointArn) {
-        SubscribeRequest request = SubscribeRequest.builder()
-                .protocol("application")
-                .endpoint(endpointArn)
-                .returnSubscriptionArn(true)
-                .topicArn(this.topicArn)
-                .build();
-
-        try {
-            SubscribeResponse response = this.snsClient.subscribe(request);
-            return response.subscriptionArn();
-        } catch (SnsException e) {
-            throw new RuntimeException("Failed to subscribe endpoint to topic", e); // Exception 추가 필요
-        }
-    }
-
-    private String getPlatformApplicationArn(String platform) {
-        if(platform.equalsIgnoreCase("ios")) {
-            return this.iosArn;
-        } else if (platform.equalsIgnoreCase("android")) {
-            return this.androidArn;
-        } else {
-            throw new IllegalStateException("Unknow platform"); // Exception 추가 필요
-        }
-    }
 
     @Override
     public String sendGeneralNotification(String message, String subject) {
