+++ src/main/java/com/locat/api/global/notification/PlatformEndpointService.java
@@ -1,0 +1,5 @@
+package com.locat.api.global.notification;
+
+public interface PlatformEndpointService {
+    String createPlatformEndpoint(String token, String platform);
+}
