+++ src/main/java/com/locat/api/global/notification/TopicSubscriptionService.java
@@ -1,0 +1,5 @@
+package com.locat.api.global.notification;
+
+public interface TopicSubscriptionService {
+    String subscribeEndpointToTopic(String endpointArn);
+}
